File clean up to fix tests
