+++ src/common/authentication.ts
@@ -1,3 +1,19 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export enum GitHubServerType {
+	None,
+	GitHubDotCom,
+	Enterprise
+}
+
+export enum AuthProvider {
+	github = 'github',
+	'github-enterprise' = 'github-enterprise'
+}
+
 export class AuthenticationError extends Error {
 	name: string;
 	stack?: string;
@@ -6,6 +22,6 @@ export class AuthenticationError extends Error {
 	}
 }
 
-export function isSamlError(e: {message?: string}): boolean {
+export function isSamlError(e: { message?: string }): boolean {
 	return !!e.message?.startsWith('Resource protected by organization SAML enforcement.');
 }
