+++ src/common/remote.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Repository } from '../api/api';
-import { AuthProvider } from '../github/credentials';
 import { getEnterpriseUri } from '../github/utils';
+import { AuthProvider, GitHubServerType } from './authentication';
 import { Protocol } from './protocol';
 
 export class Remote {
@@ -32,7 +32,7 @@ export class Remote {
 		public readonly remoteName: string,
 		public readonly url: string,
 		public readonly gitProtocol: Protocol,
-	) {}
+	) { }
 
 	equals(remote: Remote): boolean {
 		if (this.remoteName !== remote.remoteName) {
@@ -73,7 +73,7 @@ export function parseRemote(remoteName: string, url: string, originalProtocol?: 
 export function parseRepositoryRemotes(repository: Repository): Remote[] {
 	const remotes: Remote[] = [];
 	for (const r of repository.state.remotes) {
-		const urls: string[] =[];
+		const urls: string[] = [];
 		if (r.fetchUrl) {
 			urls.push(r.fetchUrl);
 		}
@@ -88,4 +88,19 @@ export function parseRepositoryRemotes(repository: Repository): Remote[] {
 		});
 	}
 	return remotes;
+}
+
+export class GitHubRemote extends Remote {
+	static remoteAsGitHub(remote: Remote, githubServerType: GitHubServerType): GitHubRemote {
+		return new GitHubRemote(remote.remoteName, remote.url, remote.gitProtocol, githubServerType);
+	}
+
+	constructor(
+		remoteName: string,
+		url: string,
+		gitProtocol: Protocol,
+		public readonly githubServerType: GitHubServerType
+	) {
+		super(remoteName, url, gitProtocol);
+	}
 }
