+++ src/github/folderRepositoryManager.ts
@@ -7,10 +7,11 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import type { Branch, Repository, UpstreamRef } from '../api/api';
 import { GitApiImpl, GitErrorCodes, RefType } from '../api/api1';
-import { GitHubManager, GitHubServerType } from '../authentication/githubServer';
+import { GitHubManager } from '../authentication/githubServer';
+import { AuthProvider, GitHubServerType } from '../common/authentication';
 import Logger from '../common/logger';
 import { Protocol, ProtocolType } from '../common/protocol';
-import { parseRepositoryRemotes, Remote } from '../common/remote';
+import { GitHubRemote, parseRepositoryRemotes, Remote } from '../common/remote';
 import { PULL_BRANCH } from '../common/settingKeys';
 import { ITelemetry } from '../common/telemetry';
 import { EventType, TimelineEvent } from '../common/timelineEvent';
@@ -21,8 +22,8 @@ import { NEVER_SHOW_PULL_NOTIFICATION, REPO_KEYS, ReposState } from '../extensio
 import { git } from '../gitProviders/gitCommands';
 import { UserCompletion, userMarkdown } from '../issues/util';
 import { OctokitCommon } from './common';
-import { AuthProvider, CredentialStore } from './credentials';
-import { GitHubRemote, GitHubRepository, ItemsData, PullRequestData, ViewerPermission } from './githubRepository';
+import { CredentialStore } from './credentials';
+import { GitHubRepository, ItemsData, PullRequestData, ViewerPermission } from './githubRepository';
 import { PullRequestState, UserResponse } from './graphql';
 import { IAccount, ILabel, IMilestone, IPullRequestsPagingOptions, PRType, RepoAccessAndMergeMethods, User } from './interface';
 import { IssueModel } from './issueModel';
