+++ src/test/github/pullRequestModel.test.ts
@@ -9,7 +9,7 @@ import { CredentialStore } from '../../github/credentials';
 import { PullRequestModel } from '../../github/pullRequestModel';
 import { GithubItemStateEnum } from '../../github/interface';
 import { Protocol } from '../../common/protocol';
-import { Remote } from '../../common/remote';
+import { GitHubRemote, Remote } from '../../common/remote';
 import { convertRESTPullRequestToRawPullRequest } from '../../github/utils';
 import { SinonSandbox, createSandbox } from 'sinon';
 import { PullRequestBuilder } from '../builders/rest/pullRequestBuilder';
@@ -18,8 +18,7 @@ import { MockGitHubRepository } from '../mocks/mockGitHubRepository';
 import { NetworkStatus } from 'apollo-client';
 import { Resource } from '../../common/resources';
 import { MockExtensionContext } from '../mocks/mockExtensionContext';
-import { GitHubRemote } from '../../github/githubRepository';
-import { GitHubServerType } from '../../authentication/githubServer';
+import { GitHubServerType } from '../../common/authentication';
 const queries = require('../../github/queries.gql');
 
 const telemetry = new MockTelemetry();
